third_party/WebKit/LayoutTests/platform/mac-mac10.10/virtual/sharedarraybuffer/fast/dom/StyleSheet/gc-styleheet-wrapper-expected.txt

Test that custom properties on stylesheets survive GC

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.firstChild.sheet.title is 'testSheet'
FAIL document.firstChild.sheet.foo should be bar (of type string). Was undefined (of type undefined).
PASS document.getElementsByTagName('link')[0].sheet.title is 'testSheet'
FAIL document.getElementsByTagName('link')[0].sheet.foo should be bar (of type string). Was undefined (of type undefined).
PASS document.getElementsByTagName('style')[0].sheet.title is 'testSheet'
PASS document.getElementsByTagName('style')[0].sheet.foo is 'bar'
PASS successfullyParsed is true

TEST COMPLETE
